test: add some tests for resource and service layer

+++ src/main/java/nl/han/oose/dea/services/PlaylistService.java
@@ -7,6 +7,7 @@ import nl.han.oose.dea.dto.incoming.PlaylistDTO;
 import nl.han.oose.dea.dto.outgoing.PlaylistResponseDTO;
 import nl.han.oose.dea.dto.incoming.TrackDTO;
 import nl.han.oose.dea.dto.outgoing.TrackResponseDTO;
+import nl.han.oose.dea.services.exceptions.TokenDoesNotExistException;
 import nl.han.oose.dea.utils.UserAuth;
 
 
@@ -29,42 +30,60 @@ public class PlaylistService extends UserAuth {
     }
 
     public PlaylistResponseDTO getPlaylists(String token) {
-        int userId = userDAO.getUserIdByToken(token);
-        return playlistDAO.getAllPlaylists(userId);
+        if (checkIfTokenExists(token)) {
+            int userId = userDAO.getUserId(token);
+            return playlistDAO.getAllPlaylists(userId);
+        } else {
+            throw new TokenDoesNotExistException("The token does not exist in the database");
+        }
     }
 
     public PlaylistResponseDTO addPlaylist(String token, PlaylistDTO p) {
-        int userId = userDAO.getUserIdByToken(token);
-        playlistDAO.addPlaylist(userId, p.getName());
-        return playlistDAO.getAllPlaylists(userId);
+        if (checkIfTokenExists(token)) {
+            int userId = userDAO.getUserId(token);
+            playlistDAO.addPlaylist(userId, p.getName());
+            return playlistDAO.getAllPlaylists(userId);
+        } else {
+            throw new TokenDoesNotExistException("The token does not exist in the database");
+        }
     }
 
     public PlaylistResponseDTO deletePlaylist(int playlistId, String token) {
-        int userId = userDAO.getUserIdByToken(token);
-        playlistDAO.deletePlaylist(playlistId);
-        return playlistDAO.getAllPlaylists(userId);
+        if (checkIfTokenExists(token)) {
+            int userId = userDAO.getUserId(token);
+            playlistDAO.deletePlaylist(playlistId);
+            return playlistDAO.getAllPlaylists(userId);
+        } else {
+            throw new TokenDoesNotExistException("The token does not exist in the database");
+        }
     }
 
     public TrackResponseDTO getAllTracksFromPlaylist(String token, int playlistId) {
         if (checkIfTokenExists(token)) {
             TrackResponseDTO tracks = new TrackResponseDTO(playlistDAO.getAllTracksById(playlistId));
             return tracks;
+        } else {
+            throw new TokenDoesNotExistException("The token does not exist in the database");
         }
-        return null;
     }
 
     public PlaylistResponseDTO updatePlaylist(int playlistId, String name, String token) {
-        int userId = userDAO.getUserIdByToken(token);
-        playlistDAO.updatePlaylist(playlistId, name);
-        return playlistDAO.getAllPlaylists(userId);
+        if (checkIfTokenExists(token)) {
+            int userId = userDAO.getUserId(token);
+            playlistDAO.updatePlaylist(playlistId, name);
+            return playlistDAO.getAllPlaylists(userId);
+        } else {
+            throw new TokenDoesNotExistException("The token does not exist in the database");
+        }
     }
 
     public TrackResponseDTO deleteTrackFromPlaylist(int playlistId, int trackId, String token) {
         if (checkIfTokenExists(token)) {
             playlistDAO.deleteTrackFromPlaylist(playlistId, trackId);
             return new TrackResponseDTO(playlistDAO.getAllTracksById(playlistId));
+        } else {
+            throw new TokenDoesNotExistException("The token does not exist in the database");
         }
-        return null;
     }
 
 
@@ -73,7 +92,8 @@ public class PlaylistService extends UserAuth {
         if (checkIfTokenExists(token)) {
             playlistDAO.addTrackToPlaylist(playlistId, track);
             return new TrackResponseDTO(playlistDAO.getAllTracksById(playlistId));
+        } else {
+            throw new TokenDoesNotExistException("The token does not exist in the database");
         }
-        return null;
     }
 }
